Improve project page
